Widen MUF integers to 64-bit
The MUF stack integer is now MUFINT (int64_t, db.h). The inst union
does not grow: it already held a double. dbrefs stay 32-bit, and
RANDOM keeps its documented 0..2^31-1 range on purpose, because real
programs normalize rolls by dividing by 2147483647.

PushInt is now a by-value push_mufint call, so all ~270 push sites
are width-correct regardless of the local's type, and the generic
push() reads a full MUFINT for PROG_INTEGER. Value-bearing paths
widened by hand: all p_math arithmetic and comparisons with
INT64 bounds on the overflow flags, ATOI and number literals via
strtoll, INTOSTR and the intostr() helper, HTOI/ITOH at full width,
the FMTSTRING %i engines (they build the format dynamically, so
-Wformat could not see them), the property integer channel end to
end (PData.val, get_property_value, db parse and print), ARRAY_SUM's
accumulator, the array intval helper, and the count/index locals in
the p_stack prims so bounds checks compare at full width before
narrowing. The array key compare is now a three-way test instead of
subtraction, which truncated and could overflow.

A compiler -Wformat sweep found 34 printf-family sites receiving the
widened value; each now uses %lld or an explicit narrowing cast
where the quantity is inherently small (variable slots, dbrefs).

Deliberate 32-bit boundaries, recorded in docs/TODO.txt: dbrefs, MPI
arithmetic, lock-expression compares, object pennies, @tune values,
and array keys. man.txt documents the new range and RANDOM's pinned
behavior.

Tests: 13 new 64-bit checks (overflow-free add and multiply, big
literal atoi round trip, 40-bit shifts both directions, big prop
round trip, fmtstring, array_sum, RANDOM range) plus the existing
suite: 120/120 pass. SSL probe still fetches https cleanly.

+++ src/inst.cpp
@@ -115,7 +115,7 @@ insttotext(struct frame *fr, int lev, struct inst *theinst, char *buffer, int bu
             }
             break;
         case PROG_INTEGER:
-            sprintf(buffer, "%d", theinst->data.number);
+            sprintf(buffer, "%lld", (long long) theinst->data.number);
             break;
         case PROG_FLOAT:
             sprintf(buffer, "%#.15g", theinst->data.fnumber);
@@ -164,42 +164,42 @@ insttotext(struct frame *fr, int lev, struct inst *theinst, char *buffer, int bu
             sprintf(buffer, "#%d", theinst->data.objref);
             break;
         case PROG_VAR:
-            sprintf(buffer, "V%d", theinst->data.number);
+            sprintf(buffer, "V%d", (int) theinst->data.number);
             break;
         case PROG_SVAR:
             if (fr) {
-                length = snprintf(buffer, buflen, "SV%d:%s", theinst->data.number, scopedvar_getname(fr, lev, theinst->data.number));
+                length = snprintf(buffer, buflen, "SV%d:%s", (int) theinst->data.number, scopedvar_getname(fr, lev, theinst->data.number));
             } else {
-                sprintf(buffer, "SV%d", theinst->data.number);
+                sprintf(buffer, "SV%d", (int) theinst->data.number);
             }
             break;
         case PROG_SVAR_AT:
         case PROG_SVAR_AT_CLEAR:
             if (fr) {
-                length = snprintf(buffer, buflen, "SV%d:%s @", theinst->data.number, scopedvar_getname(fr, lev, theinst->data.number));
+                length = snprintf(buffer, buflen, "SV%d:%s @", (int) theinst->data.number, scopedvar_getname(fr, lev, theinst->data.number));
             } else {
-                sprintf(buffer, "SV%d @", theinst->data.number);
+                sprintf(buffer, "SV%d @", (int) theinst->data.number);
             }
             break;
         case PROG_SVAR_BANG:
             if (fr) {
-                length = snprintf(buffer, buflen, "SV%d:%s !", theinst->data.number, scopedvar_getname(fr, lev, theinst->data.number));
+                length = snprintf(buffer, buflen, "SV%d:%s !", (int) theinst->data.number, scopedvar_getname(fr, lev, theinst->data.number));
             } else {
-                sprintf(buffer, "SV%d !", theinst->data.number);
+                sprintf(buffer, "SV%d !", (int) theinst->data.number);
             }
             break;
         case PROG_LVAR:
-            sprintf(buffer, "LV%d", theinst->data.number);
+            sprintf(buffer, "LV%d", (int) theinst->data.number);
             break;
         case PROG_STVAR:
-            sprintf(buffer, "STV%d", theinst->data.number);
+            sprintf(buffer, "STV%d", (int) theinst->data.number);
             break;
         case PROG_LVAR_AT:
         case PROG_LVAR_AT_CLEAR:
-            sprintf(buffer, "LV%d @", theinst->data.number);
+            sprintf(buffer, "LV%d @", (int) theinst->data.number);
             break;
         case PROG_LVAR_BANG:
-            sprintf(buffer, "LV%d !", theinst->data.number);
+            sprintf(buffer, "LV%d !", (int) theinst->data.number);
             break;
 #ifdef MUF_SOCKETS
         case PROG_SOCKET:
